requirements.txt
requests
beautifulsoup4
yfinance
anthropic
schedule
python-dotenv
